Make astrant-extensions an INTERFACE library, since it is header-only.
This may lead to an error like "Your cache is probably stale, remove the entry
astrant-extensions_LIB_DEPENDS from the cache" -- in this case, either remove
the CMakeCache file and re-run cmake, or edit the file and remove the line
starting with 'astrant-extensions_LIB_DEPENDS'.

This is only supported since CMake 3.0, while previously, astrant-extensions
supported CMake 2.8. This support is now dropped.